# backend/docs/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: Backend API for the Mihombreng controller
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Mihombreng API
  version: "1.0"
paths:
  /app/config:
    get:
      consumes:
      - application/json
      description: Get current Mihombreng application configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get application configuration
      tags:
      - Mihombreng App
    put:
      consumes:
      - application/json
      description: Update Mihombreng application configuration
      parameters:
      - description: Configuration to update
        in: body
        name: config
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update application configuration
      tags:
      - Mihombreng App
  /app/geo/ipv4:
    get:
      description: Get IPv4 address with geolocation information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Get IPv4 geolocation
      tags:
      - Mihombreng App
  /app/geo/ipv6:
    get:
      description: Get IPv6 address with geolocation information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Get IPv6 geolocation
      tags:
      - Mihombreng App
  /app/ipv4:
    get:
      description: Get current IPv4 address
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Get IPv4 address
      tags:
      - Mihombreng App
  /app/ipv6:
    get:
      description: Get current IPv6 address
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Get IPv6 address
      tags:
      - Mihombreng App
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"